requirements.txt
flask
torch
nemo_toolkit[asr]
numpy
werkzeug